Update ndk_experimental to r10 RC-3.
This updates ndk_experimental to the beta version of r10 (rc-3).  This update
is required to fix ABI compatibility with latest Android system images
(crbug.com/367378).

Note, due to the removal of __system_property_get from the exported libc.so
symbols in rc-3 (b/15704351), the libc.so files have been taken from rc-2 rather
than rc-3.

BUG=367378,354405,346626

+++ ndk_experimental/tests/run-tests.sh
@@ -36,7 +36,10 @@ NDK_BUILDTOOLS_PATH=$ROOTDIR/build/tools
 # directoy and tested separately from armeabi-v7a.  Some tests are now compiled with both
 # APP_ABI=armeabi-v7a and APP_ABI=armeabi-v7a-hard. Without _NDK_TESTING_ALL_=yes, tests
 # may fail to install due to race condition on the same libs/armeabi-v7a
-export _NDK_TESTING_ALL_=yes
+if [ -z "$_NDK_TESTING_ALL_" ]; then
+   _NDK_TESTING_ALL_=all
+fi
+export _NDK_TESTING_ALL_
 
 # The list of tests that are too long to be part of a normal run of
 # run-tests.sh. Most of these do not run properly at the moment.
@@ -61,6 +64,12 @@ RUN_TESTS_FILTERED=
 NDK_PACKAGE=
 WINE=
 CONTINUE_ON_BUILD_FAIL=
+if [ -z "$TEST_DIR" ]; then
+    TEST_DIR="/tmp/ndk-$USER/tests"
+fi
+if [ -z "$TARGET_TEST_SUBDIR" ]; then
+    TARGET_TEST_SUBDIR="ndk-tests"
+fi
 
 while [ -n "$1" ]; do
     opt="$1"
@@ -81,6 +90,9 @@ while [ -n "$1" ]; do
             ;;
         --platform=*)
             PLATFORM="$optarg"
+	    ;;
+        --test-dir=*)
+            TEST_DIR="$optarg"
             ;;
         --ndk=*)
             NDK_ROOT="$optarg"
@@ -271,7 +283,6 @@ else # !FULL_TESTS
 fi # !FULL_TESTS
 
 
-TEST_DIR="/tmp/ndk-$USER/tests"
 mkdir -p $TEST_DIR
 setup_default_log_file "$TEST_DIR/build-tests.log"
 
@@ -559,11 +570,19 @@ is_incompatible_abi ()
         local APP_ABIS=`get_build_var $PROJECT APP_ABI`
         APP_ABIS=$APP_ABIS" "
         if [ "$APP_ABIS" != "${APP_ABIS%%all*}" ] ; then
-        # replace the first "all" with all available ABIs
-          ALL_ABIS=`get_build_var $PROJECT NDK_ALL_ABIS`
-          APP_ABIS_FRONT="${APP_ABIS%%all*}"
-          APP_ABIS_BACK="${APP_ABIS#*all}"
-          APP_ABIS="${APP_ABIS_FRONT}${ALL_ABIS}${APP_ABIS_BACK}"
+        # replace "all", "all32" and "all64"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL_EXPANDED`
+          _FRONT="${APP_ABIS%%all*}"
+          _BACK="${APP_ABIS#*all}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL32_EXPANDED`
+          _FRONT="${APP_ABIS%%all32*}"
+          _BACK="${APP_ABIS#*all32}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
+          _EXPANDED=`get_build_var $PROJECT NDK_APP_ABI_ALL64_EXPANDED`
+          _FRONT="${APP_ABIS%%all64*}"
+          _BACK="${APP_ABIS#*all64}"
+          APP_ABIS="${_FRONT}${_EXPANDED}${_BACK}"
         fi
         if [ "$APP_ABIS" = "${APP_ABIS%$ABI *}" ] ; then
             echo "Skipping `basename $PROJECT`: incompatible ABI, needs $APP_ABIS"
@@ -774,7 +793,7 @@ if is_testable device; then
         local TEST=$3
         local TEST_NAME="$(basename $TEST)"
         local SRCDIR
-        local DSTDIR="$4/ndk-tests"
+        local DSTDIR="$4/$TARGET_TEST_SUBDIR"
         local SRCFILE
         local DSTFILE
         local PROGRAM
@@ -900,9 +919,11 @@ if is_testable device; then
                 dump "   ---> TEST FAILED!!"
             fi
             adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTPATH"
-            for DATA in $(ls $DATAPATHS); do
-                adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTDIR/`basename $DATA`"
-            done
+            if [ -n "$DATAPATHS" ]; then
+                for DATA in $(ls $DATAPATHS); do
+                    adb_var_shell_cmd "$DEVICE" "" "rm -f $DSTDIR/`basename $DATA`"
+                done
+            fi
         done
         # Cleanup
         adb_var_shell_cmd "$DEVICE" "" rm -r $DSTDIR
@@ -960,7 +981,7 @@ if is_testable device; then
             adb_var_shell_cmd "$DEVICE" CPU_ABI2 getprop ro.product.cpu.abi2
             CPU_ABIS="$CPU_ABI1,$CPU_ABI2"
             CPU_ABIS=$(commas_to_spaces $CPU_ABIS)
-            if [ "$_NDK_TESTING_ALL_" = "yes" ]; then
+            if [ -n "$_NDK_TESTING_ALL_" ]; then
                 if [ "$CPU_ABI1" = "armeabi-v7a" -o "$CPU_ABI2" = "armeabi-v7a" ]; then
                     CPU_ABIS="$CPU_ABIS armeabi-v7a-hard"
                 fi
